moves backend/internal/connectiondata to internal/connectiondata (#3335)
