+++ backend/services/mgmt/v1alpha1/connection-data-service/service.go
@@ -2,7 +2,7 @@ package v1alpha1_connectiondataservice
 
 import (
 	"github.com/nucleuscloud/neosync/backend/gen/go/protos/mgmt/v1alpha1/mgmtv1alpha1connect"
-	"github.com/nucleuscloud/neosync/backend/internal/connectiondata"
+	"github.com/nucleuscloud/neosync/internal/connectiondata"
 )
 
 type Service struct {
