+++ backend/services/mgmt/v1alpha1/job-service/service.go
@@ -2,10 +2,10 @@ package v1alpha1_jobservice
 
 import (
 	"github.com/nucleuscloud/neosync/backend/gen/go/protos/mgmt/v1alpha1/mgmtv1alpha1connect"
-	"github.com/nucleuscloud/neosync/backend/internal/connectiondata"
 	jobhooks "github.com/nucleuscloud/neosync/backend/internal/ee/hooks/jobs"
 	"github.com/nucleuscloud/neosync/backend/internal/userdata"
 	sql_manager "github.com/nucleuscloud/neosync/backend/pkg/sqlmanager"
+	"github.com/nucleuscloud/neosync/internal/connectiondata"
 	"github.com/nucleuscloud/neosync/internal/neosyncdb"
 	clientmanager "github.com/nucleuscloud/neosync/internal/temporal/clientmanager"
 )
