+++ internal/connectiondata/connectiondata.go
@@ -9,11 +9,11 @@ import (
 	mysql_queries "github.com/nucleuscloud/neosync/backend/gen/go/db/dbschemas/mysql"
 	pg_queries "github.com/nucleuscloud/neosync/backend/gen/go/db/dbschemas/postgresql"
 	mgmtv1alpha1 "github.com/nucleuscloud/neosync/backend/gen/go/protos/mgmt/v1alpha1"
-	neosync_gcp "github.com/nucleuscloud/neosync/backend/internal/gcp"
 	"github.com/nucleuscloud/neosync/backend/pkg/mongoconnect"
 	"github.com/nucleuscloud/neosync/backend/pkg/sqlconnect"
 	sql_manager "github.com/nucleuscloud/neosync/backend/pkg/sqlmanager"
 	aws_manager "github.com/nucleuscloud/neosync/internal/aws"
+	neosync_gcp "github.com/nucleuscloud/neosync/internal/gcp"
 	neosynctypes "github.com/nucleuscloud/neosync/internal/neosync-types"
 )
